+++ Assets/VRM/UniVRM/Editor/Meta/VRMMetaObjectEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c244f35fd2772004c9d51095f6326420
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
